new serverInfo API route
