+++ features/AMQP.feature
@@ -4,6 +4,11 @@ Feature: Test AMQP API
   Using AMQP API
 
   @usingAMQP
+  Scenario: Get server information
+    When I get server informations
+    Then I can retrieve the Kuzzle API version
+
+  @usingAMQP
   Scenario: Publish a realtime message
     When I publish a message
     Then I should receive a request id
