+++ features/MQTT.feature
@@ -4,6 +4,11 @@ Feature: Test MQTT API
   Using MQTT API
 
   @usingMQTT
+  Scenario: Get server information
+    When I get server informations
+    Then I can retrieve the Kuzzle API version
+
+  @usingMQTT
   Scenario: Publish a realtime message
     When I publish a message
     Then I should receive a request id
