+++ features/REST.feature
@@ -4,6 +4,11 @@ Feature: Test REST API
   Using REST API
 
   @usingREST
+  Scenario: Get server information
+    When I get server informations
+    Then I can retrieve the Kuzzle API version
+
+  @usingREST
   Scenario: Publish a realtime message
     When I publish a message
     Then I should receive a request id
