+++ features/STOMP.feature
@@ -4,6 +4,11 @@ Feature: Test STOMP API
   Using STOMP API
 
   @usingSTOMP
+  Scenario: Get server information
+    When I get server informations
+    Then I can retrieve the Kuzzle API version
+
+  @usingSTOMP
   Scenario: Publish a realtime message
     When I publish a message
     Then I should receive a request id
