+++ features/Websocket.feature
@@ -4,6 +4,11 @@ Feature: Test websocket API
   Using WebSocket API
 
   @usingWebsocket
+  Scenario: Get server information
+    When I get server informations
+    Then I can retrieve the Kuzzle API version
+
+  @usingWebsocket
   Scenario: Publish a realtime message
     When I publish a message
     Then I should receive a request id
