+++ features/support/apiREST.js
@@ -1,6 +1,7 @@
 var
   config = require('./config')(),
-  rp = require('request-promise');
+  rp = require('request-promise'),
+  apiVersion;
 
 var ApiREST = function () {
   this.world = null;
@@ -13,7 +14,13 @@ ApiREST.prototype.init = function (world) {
 ApiREST.prototype.disconnect = function () {};
 
 ApiREST.prototype.pathApi = function (path) {
-  return config.url + '/api/v1.0/' + path;
+  var basePath = '/api';
+
+  if (apiVersion) {
+    basePath += '/' + apiVersion;
+  }
+
+  return config.url + basePath + '/' + path;
 };
 
 ApiREST.prototype.callApi = function (options) {
@@ -256,6 +263,20 @@ ApiREST.prototype.deleteIndex = function (index) {
   };
 
   return this.callApi(options);
+};
+
+ApiREST.prototype.getServerInfo = function () {
+  var options = {
+    url: this.pathApi('_serverInfo'),
+    method: 'GET',
+    json: true
+  };
+
+  return this.callApi(options)
+    .then(res => {
+      apiVersion = res.result.serverInfo.kuzzle.api.version;
+      return res;
+    });
 };
 
 module.exports = ApiREST;
