+++ lib/api/controllers/readController.js
@@ -1,5 +1,7 @@
 var
   q = require('q'),
+  os = require('os'),
+  _ = require('lodash'),
   BadRequestError = require('../core/errors/badRequestError'),
   ResponseObject = require('../core/models/responseObject');
 
@@ -86,5 +88,84 @@ module.exports = function ReadController (kuzzle) {
     kuzzle.pluginsManager.trigger('data:listIndexes', requestObject);
 
     return kuzzle.services.list.readEngine.listIndexes(requestObject);
+  };
+
+  this.serverInfo = function (requestObject) {
+    var
+      serviceName,
+      response = {
+        kuzzle: {
+          version: kuzzle.config.version,
+          api: {
+            version: kuzzle.config.apiVersion,
+            routes: {}
+          },
+          nodeVersion: process.version,
+          memoryUsed: process.memoryUsage().rss,
+          uptime: process.uptime() + 's',
+          plugins: {},
+          system: {
+            memory: {
+              total: os.totalmem(),
+              free: os.freemem()
+            },
+            cpus: os.cpus()
+          }
+        },
+        services: {}
+      };
+
+    Object.keys(kuzzle.funnel).forEach(controller => {
+      var actionList = [];
+
+      if (typeof kuzzle.funnel[controller] === 'object') {
+        Object.keys(kuzzle.funnel[controller]).forEach(action => {
+          if (typeof kuzzle.funnel[controller][action] === 'function') {
+            actionList.push(action);
+          }
+        });
+
+        if (actionList.length > 0) {
+          response.kuzzle.api.routes[controller] = actionList;
+        }
+      }
+    });
+
+    Object.keys(kuzzle.pluginsManager.plugins).forEach(plugin => {
+      var
+        pluginInfo = kuzzle.pluginsManager.plugins[plugin],
+        p = {
+          name: pluginInfo.name,
+          version: pluginInfo.version,
+          activated: pluginInfo.activated
+        };
+
+      p.hooks = _.uniq(Object.keys(pluginInfo.object.hooks));
+
+      response.kuzzle.plugins[plugin] = p;
+    });
+
+    return Object.keys(kuzzle.services.list).reduce((previous, current) => {
+      return previous.then(serviceInfo => {
+        if (serviceInfo) {
+          response.services[serviceName] = serviceInfo;
+        }
+
+        serviceName = current;
+
+        if (kuzzle.services.list[current].getInfos) {
+          return kuzzle.services.list[current].getInfos();
+        }
+
+        return q();
+      });
+    }, q())
+    .then(serviceInfo => {
+      if (serviceInfo) {
+        response.services[serviceName] = serviceInfo;
+      }
+
+      return new ResponseObject(requestObject, {serverInfo: response});
+    });
   };
 };
