+++ lib/api/controllers/routerController.js
@@ -19,6 +19,7 @@ module.exports = function RouterController (kuzzle) {
    */
   this.initRouterHttp = function () {
     var
+      apiBase = new Router(),
       api = new Router(),
       coverage,
       routes = [
@@ -66,7 +67,17 @@ module.exports = function RouterController (kuzzle) {
     }
 
     // create and mount a new router for our API
-    this.router.use('/api/v1.0', api);
+    this.router.use('/api', apiBase);
+    this.router.use('/api/' + kuzzle.config.apiVersion, api);
+
+    /*
+     Registering the basic _serverInfo route
+     This route is also used to get Kuzzle API Version, so it isn't registered under api/<version> but
+     directly under api/
+     */
+    apiBase.get('/_serverInfo', (request, response) => {
+      executeFromRest.call(kuzzle, {controller: 'read', action: 'serverInfo'}, request, response);
+    });
 
     // create and mount a new router for plugins
     this.pluginRouter = new Router();
@@ -76,7 +87,7 @@ module.exports = function RouterController (kuzzle) {
     api.use(bodyParser.json());
 
     // Simple hello world to let know to the user that kuzzle is running
-    api.get('/', function (request, response) {
+    api.get('/', (request, response) => {
       response.writeHead('Access-Control-Allow-Origin', '*');
       response.writeHead('Access-Control-Allow-Headers', 'X-Requested-With');
       response.writeHead(200, {'Content-Type': 'application/json'});
