+++ lib/api/start.js
@@ -5,6 +5,7 @@
  */
 var
   q = require('q'),
+  packageInfo = require('../../package.json'),
   servers = require('./core/servers'),
   RequestObject = require('./core/models/requestObject'),
   HotelClerk = require('./core/hotelClerk'),
@@ -50,6 +51,8 @@ module.exports = function start (params, feature) {
   }
 
   this.config = config(params);
+  this.config.apiVersion = packageInfo.apiVersion;
+  this.config.version = packageInfo.version;
 
   this.pluginsManager = new PluginsManager(this);
   this.pluginsManager.init(this.isServer, feature.dummy);
