+++ lib/services/elasticsearch.js
@@ -38,6 +38,36 @@ module.exports = function (kuzzle, options) {
   };
 
   /**
+   * Return some basic information about this service
+   *
+   * @returns {Promise} service informations
+   */
+  this.getInfos = function () {
+    var response = {
+      type: 'elasticsearch',
+      api: this.kuzzleConfig[this.engineType].apiVersion
+    };
+
+    return this.client.info()
+      .then(res => {
+        response.version = res.version.number;
+        response.lucene = res.version.lucene_version;
+
+        return this.client.cluster.health();
+      })
+      .then(res => {
+        response.status = res.status;
+        response.nodes = res.number_of_nodes;
+        return this.client.cluster.stats({human: true});
+      })
+      .then(res => {
+        response.spaceUsed = res.indices.store.size;
+        response.nodes = res.nodes;
+        return response;
+      });
+  };
+
+  /**
    * Search documents from elasticsearch with a query
    * @param {RequestObject} requestObject
    * @returns {Promise} resolve documents matching the filter
