+++ lib/services/redis.js
@@ -51,6 +51,37 @@ module.exports = function (kuzzle, options) {
   };
 
   /**
+   * Return some basic information about this service
+   *
+   * @returns {Promise} service informations
+   */
+  this.getInfos = function () {
+    var response = {
+      type: 'redis'
+    };
+
+    return q.ninvoke(this.client, 'info').then(res => {
+      var
+        arr = res.split('\r\n'),
+        infos = {},
+        keyValuePair;
+
+      arr.forEach(item => {
+        if (item.length > 0 && !item.startsWith('#')) {
+          keyValuePair = item.split(':');
+          infos[keyValuePair[0]] = keyValuePair[1];
+        }
+      });
+
+      response.version = infos.redis_version;
+      response.mode = infos.redis_mode;
+      response.memoryUsed = infos.used_memory_human;
+      response.memoryPeak = infos.used_memory_peak_human;
+      return response;
+    });
+  };
+
+  /**
    * Add one or multiple value to a key
    *
    * @param {String} key
