+++ package.json
@@ -2,6 +2,7 @@
   "name": "kuzzle",
   "author": "The Kuzzle Team <support@kuzzle.io>",
   "version": "0.7.2",
+  "apiVersion": "1.0",
   "description": "Kuzzle is an open-source solution that handles all the data management through a secured API, with a large choice of protocols.",
   "main": "./lib/index.js",
   "bin": {
